move the analytics settings code to a dedicated file
